created constants directory
