+++ lua/configs/cmp-opts.lua
@@ -1,21 +1,13 @@
 ---@see CMP https://github.com/hrsh7th/nvim-cmp
 local cmp = require("cmp")
-local icons = require("configs.icons").cmp
-local border = require("utils").border
-
--- default fields order i.e completion word + item.kind + item.kind icons
-local formatting_style = {
-	fields = { "abbr", "kind", "menu" },
-	format = function(_, item)
-		local icon = icons[item.kind] or ""
-		item.kind = string.format("%s %s", (icon .. " ") or icon, item.kind or "")
-		return item
-	end,
-}
+local border = require("utils.ui").border
 
 return {
-	formatting = formatting_style,
 	completion = { completeopt = "menu,menuone" },
+	formatting = {
+		fields = { "abbr", "kind", "menu" },
+		format = require("utils").cmp_format,
+	},
 	sources = {
 		{ name = "nvim_lsp" },
 		{ name = "luasnip" },
@@ -23,6 +15,20 @@ return {
 		{ name = "nvim_lua" },
 		{ name = "path" },
 		{ name = "npm", keyword_length = 4 },
+	},
+	snippet = {
+		expand = function(args)
+			require("luasnip").lsp_expand(args.body)
+		end,
+	},
+	window = {
+		documentation = { border = border("CmpDocBorder"), winhighlight = "Normal:CmpDoc" },
+		completion = {
+			border = border("CmpCompBorder"),
+			scrollbar = false,
+			side_padding = 1,
+			winhighlight = "Normal:Normal,FloatBorder:FloatBorder,CursorLine:Visual,Search:None",
+		},
 	},
 	mapping = {
 		["<CR>"] = cmp.mapping.confirm({ behavior = cmp.ConfirmBehavior.Insert, select = true }),
@@ -50,19 +56,5 @@ return {
 				fallback()
 			end
 		end, { "i", "s" }),
-	},
-	snippet = {
-		expand = function(args)
-			require("luasnip").lsp_expand(args.body)
-		end,
-	},
-	window = {
-		documentation = { border = border("CmpDocBorder"), winhighlight = "Normal:CmpDoc" },
-		completion = {
-			border = border("CmpCompBorder"),
-			scrollbar = false,
-			side_padding = 1,
-			winhighlight = "Normal:Normal,FloatBorder:FloatBorder,CursorLine:Visual,Search:None",
-		},
 	},
 }
