+++ lua/configs/gitsigns-opts.lua
@@ -66,9 +66,8 @@ return {
 	current_line_blame_formatter = "<author>, <author_time:%Y-%m-%d> - <summary>",
 	sign_priority = 6,
 	update_debounce = 100,
-	status_formatter = nil, -- Use default
 	max_file_length = 40000, -- Disable if file is longer than this (in lines)
-	-- yadm = { enable = false },
+	status_formatter = nil, -- Use default
 	current_line_blame_opts = {
 		delay = 1000,
 		virt_text = true,
