+++ lua/configs/lspconfig/clangd.lua
@@ -1,11 +1,15 @@
-return {
-	clangd = {
-		fallbackFlags = { "-std=c++20" },
-		InlayHints = {
-			Designators = true,
-			Enabled = true,
-			ParameterNames = true,
-			DeducedTypes = true,
-		},
-	},
+local M = {}
+
+local InlayHints = {
+	Enabled = true,
+	Designators = true,
+	DeducedTypes = true,
+	ParameterNames = true,
 }
+
+M.clangd = {
+	fallbackFlags = { "-std=c++20" },
+	InlayHints = InlayHints,
+}
+
+return M
