+++ lua/configs/lspconfig/cssls.lua
@@ -1,5 +1,7 @@
+local opts = { validate = true, lint = { unknownAtRules = "ignore" } }
+
 return {
-	css = { validate = true, lint = { unknownAtRules = "ignore" } },
-	scss = { validate = true, lint = { unknownAtRules = "ignore" } },
-	less = { validate = true, lint = { unknownAtRules = "ignore" } },
+	css = opts,
+	scss = opts,
+	less = opts,
 }
