+++ lua/configs/lspconfig/gopls.lua
@@ -1,13 +1,17 @@
-return {
-	gopls = {
-		hints = {
-			constantValues = true,
-			parameterNames = true,
-			rangeVariableTypes = true,
-			assignVariableTypes = true,
-			compositeLiteralTypes = true,
-			functionTypeParameters = true,
-			compositeLiteralFields = true,
-		},
-	},
+local M = {}
+
+local hints = {
+	constantValues = true,
+	parameterNames = true,
+	rangeVariableTypes = true,
+	assignVariableTypes = true,
+	compositeLiteralTypes = true,
+	functionTypeParameters = true,
+	compositeLiteralFields = true,
 }
+
+M.gopls = {
+	hints = hints,
+}
+
+return M
