+++ lua/configs/lspconfig/lua_ls.lua
@@ -1,15 +1,18 @@
-return {
-	Lua = {
-		diagnostics = { globals = { "vim" } },
-		hint = { enable = true },
-		workspace = {
-			maxPreload = 100000,
-			preloadFileSize = 10000,
-			library = {
-				[vim.fn.expand("$VIMRUNTIME/lua")] = true,
-				[vim.fn.expand("$VIMRUNTIME/lua/vim/lsp")] = true,
-				[vim.fn.stdpath("data") .. "/lazy/lazy.nvim/lua/lazy"] = true,
-			},
-		},
+local M = {}
+local library = {
+	[vim.fn.expand("$VIMRUNTIME/lua")] = true,
+	[vim.fn.expand("$VIMRUNTIME/lua/vim/lsp")] = true,
+	[vim.fn.stdpath("data") .. "/lazy/lazy.nvim/lua/lazy"] = true,
+}
+
+M.Lua = {
+	hint = { enable = true },
+	diagnostics = { globals = { "vim" } },
+	workspace = {
+		library = library,
+		maxPreload = 100000,
+		preloadFileSize = 10000,
 	},
 }
+
+return M
