+++ lua/constants/colors.lua
@@ -1,0 +1,26 @@
+local M = {}
+
+---@see CratesDocumentation https://github.com/Saecki/crates.nvim
+M.PACKAGE_INFO_COLORS = {
+	up_to_date = "#9CDAE9",
+	outdated = "#F5E3B5",
+}
+
+M.DASHBOARD_COLORS = {
+	"#F5E0DC",
+	"#F2CDCD",
+	"#F5C2E7",
+	"#CBA6F7",
+	"#F38BA8",
+	"#EBA0AC",
+	"#FAB387",
+	"#F9E2AF",
+	"#A6E3A1",
+	"#94E2D5",
+	"#89DCEB",
+	"#74C7EC",
+	"#89B4FA",
+	"#B4BEFE",
+}
+
+return M
