+++ lua/constants/formatters.lua
@@ -1,0 +1,19 @@
+--- @see Formatters https://github.com/stevearc/conform.nvim?tab=readme-ov-file#formatters
+return {
+	go = { "gofmt" },
+	sh = { "shfmt" },
+	zsh = { "shfmt" },
+	lua = { "stylua" },
+	toml = { "taplo" },
+	yaml = { "yamlfmt" },
+	zshrc = { "shfmt" },
+	markdown = { "markdownlint" },
+	css = { "prettier" },
+	html = { "prettier" },
+	astro = { "prettier" },
+	svelte = { "prettier" },
+	javascript = { "prettier" },
+	typescript = { "prettier" },
+	javascriptreact = { "prettier" },
+	typescriptreact = { "prettier" },
+}
