+++ lua/constants/ft.lua
@@ -1,0 +1,41 @@
+local M = {}
+
+M.WEB_FILETYPES = {
+	"hbs",
+	"jsx",
+	"php",
+	"tsx",
+	"vue",
+	"xml",
+	"html",
+	"astro",
+	"templ",
+	"svelte",
+	"glimmer",
+	"markdown",
+	"rescript",
+	"handlebars",
+	"javascript",
+	"typescript",
+	"javascriptreact",
+	"typescriptreact",
+}
+
+M.EMMET_FILETYPES = {
+	"css",
+	"pug",
+	"vue",
+	"html",
+	"less",
+	"sass",
+	"scss",
+	"astro",
+	"eruby",
+	"svelte",
+	"htmldjango",
+	"htmlangular",
+	"javascriptreact",
+	"typescriptreact",
+}
+
+return M
