+++ lua/constants/leader.lua
@@ -1,0 +1,1 @@
+return " "
