+++ lua/constants/linters.lua
@@ -1,0 +1,12 @@
+---@see Linters https://github.com/mfussenegger/nvim-lint?tab=readme-ov-file#available-linters
+return {
+	sh = { "shellcheck" },
+	zsh = { "zsh" },
+	css = { "stylelint" },
+	html = { "htmlhint" },
+	json = { "jsonlint" },
+	-- yaml = { "yamllint" },
+	markdown = { "markdownlint" },
+	javascript = { "eslint" },
+	typescript = { "eslint" },
+}
