+++ lua/constants/plugins.lua
@@ -1,0 +1,8 @@
+--- directories to lazy load plugins from
+return {
+	"ui",
+	"lang",
+	"editor",
+	"debug",
+	"test",
+}
