+++ lua/constants/servers.lua
@@ -1,0 +1,22 @@
+---@see Servers https://github.com/neovim/nvim-lspconfig/blob/master/doc/server_configurations.md
+return {
+	"marksman",
+	"mdx_analyzer",
+	-- web
+	"html",
+	"htmx",
+	"astro",
+	"graphql",
+	"svelte",
+	"eslint",
+	"cssls",
+	"cssmodules_ls",
+	-- data
+	"sqlls",
+	"taplo", --- https://taplo.tamasfe.dev/
+	-- systems
+	"dockerls",
+	"clangd",
+	"vimls",
+	"cmake",
+}
