+++ lua/plugins/editor/lint.lua
@@ -1,24 +1,13 @@
+local callback = function()
+	require("lint").try_lint()
+end
+
 return {
 	"mfussenegger/nvim-lint",
 	init = function()
-		vim.api.nvim_create_autocmd({ "BufWritePost" }, {
-			callback = function()
-				require("lint").try_lint()
-			end,
-		})
+		vim.api.nvim_create_autocmd({ "BufWritePost" }, { callback = callback })
 	end,
 	config = function()
-		--- https://github.com/mfussenegger/nvim-lint?tab=readme-ov-file#available-linters
-		require("lint").linters_by_ft = {
-			sh = { "shellcheck" },
-			zsh = { "zsh" },
-			css = { "stylelint" },
-			html = { "htmlhint" },
-			json = { "jsonlint" },
-			-- yaml = { "yamllint" },
-			markdown = { "markdownlint" },
-			javascript = { "eslint" },
-			typescript = { "eslint" },
-		}
+		require("lint").linters_by_ft = require("constants.linters")
 	end,
 }
