+++ lua/plugins/editor/macros.lua
@@ -2,7 +2,5 @@ return {
 	"ecthelionvi/NeoComposer.nvim",
 	dependencies = { "kkharji/sqlite.lua" },
 	config = true,
-	keys = {
-		{ "<leader>mm", "<cmd>EditMacros<CR>", desc = "Edit macros" },
-	},
+	keys = { { "<leader>mm", "<cmd>EditMacros<CR>", desc = "Edit macros" } },
 }
