+++ lua/plugins/leetcode.lua
@@ -1,24 +1,25 @@
 local openURL = require("utils").openURL
 
+local keys = {
+	{ "<leader>le", "<cmd>Leet<cr>", desc = "Open leet" },
+	{ "<leader>lr", "<cmd>Leet run<cr>", desc = "Run leet" },
+	{ "<leader>lp", "<cmd>Leet open<cr>", desc = "Open leet url" },
+	{ "<leader>lc", "<cmd>Leet lang<cr>", desc = "Open leet language picker" },
+	{ "<leader>li", "<cmd>Leet info<cr>", desc = "Get info for leet question" },
+	{
+		"<leader>lh",
+		function()
+			openURL("https://github.com/kawre/leetcode.nvim?tab=readme-ov-file")
+		end,
+		desc = "Get info for leet question",
+	},
+	--
+}
 return {
 	"kawre/leetcode.nvim",
 	config = true,
 	build = ":TSUpdate html",
-	keys = {
-		{ "<leader>le", "<cmd>Leet<cr>", desc = "Open leet" },
-		{ "<leader>lr", "<cmd>Leet run<cr>", desc = "Run leet" },
-		{ "<leader>lp", "<cmd>Leet open<cr>", desc = "Open leet url" },
-		{ "<leader>lc", "<cmd>Leet lang<cr>", desc = "Open leet language picker" },
-		{ "<leader>li", "<cmd>Leet info<cr>", desc = "Get info for leet question" },
-		{
-			"<leader>lh",
-			function()
-				openURL("https://github.com/kawre/leetcode.nvim?tab=readme-ov-file")
-			end,
-			desc = "Get info for leet question",
-		},
-		--
-	},
+	keys = keys,
 	dependencies = {
 		"nvim-telescope/telescope.nvim",
 		"nvim-lua/plenary.nvim", -- required by telescope
