+++ lua/srandle/lazy.lua
@@ -1,21 +1,20 @@
-local lazypath = vim.fn.stdpath("data") .. "/lazy/lazy.nvim"
+local ENABLED_PLUGINS = require("constants.plugins")
+local LAZY_DIR = require("constants.dir").LAZY_DIR
 local plugins = require("utils.expand").expand_plugins
 local icons = require("configs.icons").lazy
 
-if not vim.loop.fs_stat(lazypath) then
+vim.opt.rtp:prepend(LAZY_DIR)
+if not vim.loop.fs_stat(LAZY_DIR) then
 	vim.fn.system({
 		"git",
 		"clone",
 		"--filter=blob:none",
 		"https://github.com/folke/lazy.nvim.git",
 		"--branch=stable", -- latest stable release
-		lazypath,
+		LAZY_DIR,
 	})
 end
 
-vim.opt.rtp:prepend(lazypath)
-
-local enabled_plugins = { "ui", "lang", "editor", "debug", "test" }
 local lazy_options = {
 	defaults = { lazy = true },
 	install = { colorscheme = { "catpuccin" } },
@@ -23,4 +22,4 @@ local lazy_options = {
 	{ icons = icons },
 }
 
-require("lazy").setup(plugins(enabled_plugins), lazy_options)
+require("lazy").setup(plugins(ENABLED_PLUGINS), lazy_options)
