--- lua/srandle/leader.lua
@@ -1,1 +1,0 @@
-vim.g.mapleader = " "
