+++ lua/utils/dashboard.lua
@@ -3,24 +3,9 @@ local M = {}
 M.expand_actions = require("utils.expand").expand_actions
 
 M.color_dashboard = function()
-	local dashboard_colors = {
-		"#f5e0dc",
-		"#f2cdcd",
-		"#f5c2e7",
-		"#cba6f7",
-		"#f38ba8",
-		"#eba0ac",
-		"#fab387",
-		"#f9e2af",
-		"#a6e3a1",
-		"#94e2d5",
-		"#89dceb",
-		"#74c7ec",
-		"#89b4fa",
-		"#b4befe",
-	}
+	local DASHBOARD_COLORS = require("constants.colors").DASHBOARD_COLORS
 	math.randomseed(os.time())
-	local selected = dashboard_colors[math.random(#dashboard_colors)]
+	local selected = DASHBOARD_COLORS[math.random(#DASHBOARD_COLORS)]
 	vim.cmd("hi DashboardHeader guibg=none guifg=" .. selected)
 end
