--- lua/utils/globals.lua
@@ -1,24 +1,0 @@
-local M = {}
-
-M.web_filetypes = {
-	"html",
-	"javascript",
-	"typescript",
-	"javascriptreact",
-	"typescriptreact",
-	"svelte",
-	"templ",
-	"vue",
-	"tsx",
-	"jsx",
-	"rescript",
-	"xml",
-	"php",
-	"markdown",
-	"astro",
-	"glimmer",
-	"handlebars",
-	"hbs",
-}
-
-return M
